Устранение дребезга
