+++ js/constants.js
@@ -1,4 +1,5 @@
 const ARRAY_USERS_COUNT = 10;
+const RERENDER_DELAY = 400;
 
 const TOKYO = {
   lat: 35.678,
@@ -13,4 +14,4 @@ const APARTMENT_OPTION = {
   'palace': '10000',
 };
 
-export {ARRAY_USERS_COUNT, TOKYO, APARTMENT_OPTION};
+export {ARRAY_USERS_COUNT, TOKYO, APARTMENT_OPTION, RERENDER_DELAY};
