+++ js/main.js
@@ -1,6 +1,8 @@
 import './ad-form-validate.js';
 import './slider.js';
 
+import {debounce} from './util.js';
+import {RERENDER_DELAY} from './constants.js';
 import {advertisementFilter} from './filter.js';
 import {serverAdvertisementPinMap} from './map.js';
 import {getDataLoadingErrorMessage} from './data.js';
@@ -8,5 +10,10 @@ import {getData} from './api.js';
 
 getData((data) => {
   serverAdvertisementPinMap(data);
-  advertisementFilter(data, serverAdvertisementPinMap);
+  advertisementFilter(
+    data,
+    debounce(
+      serverAdvertisementPinMap,
+      RERENDER_DELAY,
+    ));
 }, getDataLoadingErrorMessage);
