+++ js/util.js
@@ -1,0 +1,66 @@
+const getPlaceText = (place) => {
+  switch (place) {
+    case 'palace':
+      return 'Дворец';
+    case 'flat':
+      return 'Квартира';
+    case 'house':
+      return 'Дом';
+    case 'bungalow':
+      return 'Бунгало';
+    case 'hotel':
+      return 'Отель';
+    default:
+      return 'Не выбран тип жилья!';
+  }
+};
+
+const getRoomsText = (rooms) => {
+  switch (rooms) {
+    case 1:
+      return 'комната';
+    case 2:
+    case 3:
+    case 4:
+      return 'комнаты';
+    case 5:
+    case 6:
+    case 7:
+    case 8:
+    case 9:
+    case 10:
+      return 'комнат';
+    default:
+      return '';
+  }
+};
+
+const getGuestsText = (guests) => {
+  let word;
+  if (guests === 1) {
+    word = 'гостя';
+  } else {
+    word = 'гостей';
+  }
+  return word;
+};
+
+function debounce (callback, timeoutDelay = 500) {
+  // Используем замыкания, чтобы id таймаута у нас навсегда приклеился
+  // к возвращаемой функции с setTimeout, тогда мы его сможем перезаписывать
+  let timeoutId;
+
+  return (...rest) => {
+    // Перед каждым новым вызовом удаляем предыдущий таймаут,
+    // чтобы они не накапливались
+    clearTimeout(timeoutId);
+
+    // Затем устанавливаем новый таймаут с вызовом колбэка на ту же задержку
+    timeoutId = setTimeout(() => callback.apply(this, rest), timeoutDelay);
+
+    // Таким образом цикл «поставить таймаут - удалить таймаут» будет выполняться,
+    // пока действие совершается чаще, чем переданная задержка timeoutDelay
+  };
+}
+
+export {getPlaceText, getRoomsText, getGuestsText, debounce};
